+++ account/src/main/java/bisq/account/protocol_type/ProtocolType.java
@@ -20,6 +20,10 @@ package bisq.account.protocol_type;
 import bisq.common.proto.ProtoEnum;
 import bisq.common.proto.UnresolvableProtobufMessageException;
 
+// TODO does not fit right into the account module. But dependency graph does not provide a lower level module
+// where it would fit in. protocol module is rather high level and we use protocol type in the POC code in 
+// the account, offer and contract modules. So leave it for now here. Maybe we need another low level exchange domain 
+// specific module where it would fit in.
 public interface ProtocolType extends ProtoEnum {
     static ProtocolType fromProto(bisq.account.protobuf.ProtocolType proto) {
         switch (proto.getMessageCase()) {
