+++ desktop/src/main/java/bisq/desktop/primary/overlay/bisq_easy/create_offer/payment_method/CreateOfferPaymentMethodController.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.primary.overlay.bisq_easy.create_offer.method;
+package bisq.desktop.primary.overlay.bisq_easy.create_offer.payment_method;
 
 import bisq.account.payment_method.FiatPaymentRail;
 import bisq.account.payment_method.FiatPaymentUtil;
