+++ desktop/src/main/java/bisq/desktop/primary/overlay/bisq_easy/create_offer/payment_method/CreateOfferPaymentMethodView.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.primary.overlay.bisq_easy.create_offer.method;
+package bisq.desktop.primary.overlay.bisq_easy.create_offer.payment_method;
 
 import bisq.desktop.common.utils.ImageUtil;
 import bisq.desktop.common.view.View;
