+++ desktop/src/main/java/bisq/desktop/primary/overlay/bisq_easy/take_offer/payment_method/TakeOfferPaymentController.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.primary.overlay.bisq_easy.take_offer.payment;
+package bisq.desktop.primary.overlay.bisq_easy.take_offer.payment_method;
 
 import bisq.application.DefaultApplicationService;
 import bisq.desktop.common.view.Controller;
